Update constants and move telemetry to different file
